requirements.txt
Flask==2.3.2
Flask-Limiter==3.3.1
python-dotenv==1.0.0
pytube==15.0.0
ffmpeg-python==0.4.7
asyncio==3.4.3
gunicorn==20.1.0
